Generic Service Extentions
