--- src/Identity.Dapper.PostgreSQL/Extensions/ServiceCollectionExtensions.cs
@@ -1,117 +1,0 @@
-﻿using Identity.Dapper.Connections;
-using Identity.Dapper.Cryptography;
-using Identity.Dapper.Entities;
-using Identity.Dapper.Models;
-using Identity.Dapper.PostgreSQL.Connections;
-using Identity.Dapper.PostgreSQL.Models;
-using Identity.Dapper.Repositories;
-using Identity.Dapper.Repositories.Contracts;
-using Identity.Dapper.Stores;
-using Identity.Dapper.UnitOfWork.Contracts;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-
-namespace Identity.Dapper.PostgreSQL
-{
-    public static class ServiceCollectionExtensions
-    {
-        public static IdentityBuilder AddDapperIdentityForPostgreSql(this IdentityBuilder builder)
-        {
-            builder.Services.AddSingleton<SqlConfiguration, PostgreSqlConfiguration>();
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType);
-
-            return builder;
-        }
-
-        public static IdentityBuilder AddDapperIdentityForPostgreSql(this IdentityBuilder builder, PostgreSqlConfiguration configurationOverride)
-        {
-            builder.Services.AddSingleton<SqlConfiguration>(configurationOverride);
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType);
-
-            return builder;
-        }
-
-        public static IdentityBuilder AddDapperIdentityForPostgreSql<TKey>(this IdentityBuilder builder)
-        {
-            builder.Services.AddSingleton<SqlConfiguration, PostgreSqlConfiguration>();
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType, typeof(TKey));
-
-            return builder;
-        }
-
-        public static IdentityBuilder AddDapperIdentityForPostgreSql<TKey>(this IdentityBuilder builder, PostgreSqlConfiguration configurationOverride)
-        {
-            builder.Services.AddSingleton<SqlConfiguration>(configurationOverride);
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType, typeof(TKey));
-
-            return builder;
-        }
-
-        public static IdentityBuilder AddDapperIdentityForPostgreSql<TKey, TUserRole, TRoleClaim>(this IdentityBuilder builder)
-        {
-            builder.Services.AddSingleton<SqlConfiguration, PostgreSqlConfiguration>();
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType, typeof(TKey), typeof(TUserRole), typeof(TRoleClaim));
-
-            return builder;
-        }
-
-        public static IdentityBuilder AddDapperIdentityForPostgreSql<TKey, TUserRole, TRoleClaim>(this IdentityBuilder builder, PostgreSqlConfiguration configurationOverride)
-        {
-            builder.Services.AddSingleton<SqlConfiguration>(configurationOverride);
-            builder.Services.AddScoped<IUnitOfWork, UnitOfWork.UnitOfWork>();
-
-            AddStores(builder.Services, builder.UserType, builder.RoleType, typeof(TKey), typeof(TUserRole), typeof(TRoleClaim));
-
-            return builder;
-        }
-
-        private static void AddStores(IServiceCollection services, Type userType, Type roleType, Type keyType = null, Type userRoleType = null, Type roleClaimType = null, Type userClaimType = null, Type userLoginType = null)
-        {
-            Type userStoreType;
-            Type roleStoreType;
-            keyType = keyType ?? typeof(int);
-            userRoleType = userRoleType ?? typeof(DapperIdentityUserRole<>).MakeGenericType(keyType);
-            roleClaimType = roleClaimType ?? typeof(DapperIdentityRoleClaim<>).MakeGenericType(keyType);
-            userClaimType = userClaimType ?? typeof(DapperIdentityUserClaim<>).MakeGenericType(keyType);
-            userLoginType = userLoginType ?? typeof(DapperIdentityUserLogin<>).MakeGenericType(keyType);
-
-            userStoreType = typeof(DapperUserStore<,,,,,,>).MakeGenericType(userType, keyType, userRoleType, roleClaimType,
-                                                                           userClaimType, userLoginType, roleType);
-            roleStoreType = typeof(DapperRoleStore<,,,>).MakeGenericType(roleType, keyType, userRoleType, roleClaimType);
-
-            services.AddScoped(typeof(IRoleRepository<,,,>).MakeGenericType(roleType, keyType, userRoleType, roleClaimType),
-                               typeof(RoleRepository<,,,>).MakeGenericType(roleType, keyType, userRoleType, roleClaimType));
-
-            services.AddScoped(typeof(IUserRepository<,,,,,,>).MakeGenericType(userType, keyType, userRoleType,
-                                                                              roleClaimType, userClaimType,
-                                                                              userLoginType, roleType),
-                               typeof(UserRepository<,,,,,,>).MakeGenericType(userType, keyType, userRoleType,
-                                                                             roleClaimType, userClaimType,
-                                                                             userLoginType, roleType));
-
-            services.AddScoped(typeof(IUserStore<>).MakeGenericType(userType), userStoreType);
-            services.AddScoped(typeof(IRoleStore<>).MakeGenericType(roleType), roleStoreType);
-        }
-
-        public static IServiceCollection ConfigureDapperPostgreSqlConnectionProvider(this IServiceCollection services, IConfigurationSection configuration)
-        {
-            services.Configure<ConnectionProviderOptions>(configuration);
-
-            services.AddScoped<IConnectionProvider, PostgreSqlConnectionProvider>();
-
-            return services;
-        }
-    }
-}
